requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]>=0.31.1
gunicorn>=22.0.0

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic>=2,<3

# ── Azure Identity (Managed Identity / DefaultAzureCredential) ────────────────
# Used by: SearchClient (retrieval), AzureOpenAI embeddings, AzureOpenAIChatClient
azure-identity>=1.17.0

# ── Azure AI Search ───────────────────────────────────────────────────────────
# 11.6.0b3 required for VectorizedQuery + semantic reranker support
azure-search-documents==11.6.0b3

# ── Azure OpenAI ──────────────────────────────────────────────────────────────
# openai>=2.0.0 required for agent-framework-core compatibility
# azure_ad_token_provider supported on both AzureOpenAI and AsyncAzureOpenAI
openai>=2.0.0

# ── Microsoft Agent Framework SDK ─────────────────────────────────────────────
# Provides: AzureOpenAIChatClient, client.as_agent(), InMemoryHistoryProvider,
#           BaseContextProvider, SessionContext, AgentSession
agent-framework-core==1.0.0rc3

# ── HTTP client (used by openai and Agent Framework SDK internally) ────────────
httpx>=0.27.1
